pre made models for app (different settings)
These models are all made using the template files but differ in settings (indicated in their titles). This means their accuracy is different. For example the unlimited depth model has a R2 of 1. in the training set but a bad score on the test sets. The models with fewer depth (meaning less decision tree levels) are more accurate on the test sets since their is fewer overfitting.